refactor(ai-workflow-builder): Move triage agent to code-builder/
The triage agent is the entry point to the builder system, so it
logically belongs in code-builder/ rather than agents/. The agents/
folder now only retains the deprecated multi-agent pieces (planner,
responder, supervisor).
