+++ packages/@n8n/ai-workflow-builder.ee/src/code-builder/tools/ask-assistant.tool.ts
