+++ packages/@n8n/ai-workflow-builder.ee/src/code-builder/tools/build-workflow.tool.ts
